+++ src/App.jsx
@@ -6,7 +6,6 @@ import { getSelectors } from './selectors';
 // View Components
 import Root from './views/Root';
 import UserHome from './views/UserHome';
-import LoginView from './views/Login';
 
 const App = () => {
 	const { isAuthenticated } = getSelectors().user;
@@ -15,7 +14,6 @@ const App = () => {
 			<Container maxWidth="lg">
 				<Route exact path="/">
 					{isAuthenticated ? <UserHome /> : <Root /> }
-					<LoginView />
 				</Route>
 			</Container>
 		</div>
